Renaming files according to coding stile guide.
