+++ sharc/propagation/propagation_abg.py
